Broke out 'ItemPointer' from types.h and put it in its own files in 'common'. Updated everybody that needs this to include the new header.
