+++ src/include/logging/backend_logger.h
@@ -10,19 +10,19 @@
 //
 //===----------------------------------------------------------------------===//
 
-
 #pragma once
 
-#include <vector>
-#include <mutex>
 #include <condition_variable>
+#include <mutex>
+#include <vector>
 
-#include "type/types.h"
-#include "logging/logger.h"
-#include "logging/log_record.h"
-#include "logging/log_buffer.h"
+#include "common/item_pointer.h"
 #include "common/platform.h"
 #include "logging/circular_buffer_pool.h"
+#include "logging/log_buffer.h"
+#include "logging/log_record.h"
+#include "logging/logger.h"
+#include "type/types.h"
 
 namespace peloton {
 namespace logging {
