+++ src/include/storage/tile_group.h
@@ -10,20 +10,20 @@
 //
 //===----------------------------------------------------------------------===//
 
-
 #pragma once
 
-#include <map>
 #include <atomic>
-#include <vector>
-#include <mutex>
+#include <map>
 #include <memory>
+#include <mutex>
+#include <vector>
 
+#include "common/item_pointer.h"
+#include "common/printable.h"
+#include "planner/project_info.h"
+#include "type/abstract_pool.h"
 #include "type/types.h"
 #include "type/value.h"
-#include "common/printable.h"
-#include "type/abstract_pool.h"
-#include "planner/project_info.h"
 
 namespace peloton {
 
@@ -164,7 +164,7 @@ class TileGroup : public Printable {
   // Sets the tile id and column id w.r.t that tile corresponding to
   // the specified tile group column id.
   inline void LocateTileAndColumn(oid_t column_offset, oid_t &tile_offset,
-                           oid_t &tile_column_offset) {
+                                  oid_t &tile_column_offset) {
     PL_ASSERT(column_map.count(column_offset) != 0);
     // get the entry in the column map
     auto entry = column_map.at(column_offset);
