+++ src/main/sdbench/sdbench_loader.cpp
@@ -23,6 +23,7 @@
 
 #include "catalog/manager.h"
 #include "catalog/schema.h"
+#include "common/item_pointer.h"
 #include "common/logger.h"
 #include "common/macros.h"
 #include "concurrency/transaction.h"
@@ -52,9 +53,9 @@ void CreateTable() {
   std::vector<catalog::Column> columns;
 
   for (oid_t col_itr = 0; col_itr < col_count; col_itr++) {
-    auto column =
-        catalog::Column(type::Type::INTEGER, type::Type::GetTypeSize(type::Type::INTEGER),
-                        "" + std::to_string(col_itr), is_inlined);
+    auto column = catalog::Column(type::Type::INTEGER,
+                                  type::Type::GetTypeSize(type::Type::INTEGER),
+                                  "" + std::to_string(col_itr), is_inlined);
 
     columns.push_back(column);
   }
@@ -113,7 +114,6 @@ void LoadTable() {
   } else {
     LOG_TRACE("commit failed");
   }
-
 }
 
 void CreateAndLoadTable(LayoutType layout_type) {
