+++ src/type/types.cpp
@@ -24,8 +24,6 @@
 
 namespace peloton {
 
-ItemPointer INVALID_ITEMPOINTER;
-
 FileHandle INVALID_FILE_HANDLE;
 
 // WARNING: It will limit scalability if tuples per tile group is too small,
@@ -321,14 +319,6 @@ bool HexDecodeToBinary(unsigned char* bufferdst, const char* hexString) {
     bufferdst[i] = static_cast<unsigned char>(result);
   }
   return true;
-}
-
-bool AtomicUpdateItemPointer(ItemPointer* src_ptr, const ItemPointer& value) {
-  PL_ASSERT(sizeof(ItemPointer) == sizeof(int64_t));
-  int64_t* cast_src_ptr = reinterpret_cast<int64_t*>((void*)src_ptr);
-  int64_t* cast_value_ptr = reinterpret_cast<int64_t*>((void*)&value);
-  return __sync_bool_compare_and_swap(cast_src_ptr, *cast_src_ptr,
-                                      *cast_value_ptr);
 }
 
 //===--------------------------------------------------------------------===//
